+++ project-model/src/main/java/com/walmartlabs/concord/project/yaml/model/YamlCall.java
@@ -2,23 +2,32 @@ package com.walmartlabs.concord.project.yaml.model;
 
 import com.fasterxml.jackson.core.JsonLocation;
 
+import java.util.Map;
+
 public class YamlCall extends YamlStep {
 
-    private final String proc;
+    private final String key;
+    private final Map<String, Object> options;
 
-    public YamlCall(JsonLocation location, String proc) {
+    public YamlCall(JsonLocation location, String key, Map<String, Object> options) {
         super(location);
-        this.proc = proc;
+        this.key = key;
+        this.options = options;
     }
 
-    public String getProc() {
-        return proc;
+    public String getKey() {
+        return key;
+    }
+
+    public Map<String, Object> getOptions() {
+        return options;
     }
 
     @Override
     public String toString() {
         return "YamlCall{" +
-                "proc='" + proc + '\'' +
+                "key='" + key + '\'' +
+                ", options=" + options +
                 '}';
     }
 }